move traits & types to appropriate locations
